test: convert backend client tests from fixture files to inline payloads
