--- tests/required/client/modelserver/fixtures/sglang/chat.json
@@ -1,26 +1,0 @@
-{
-  "id": "5f2f8f6a1b3c4d5e8f7a6b5c4d3e2f1a",
-  "object": "chat.completion",
-  "created": 1750000000,
-  "model": "meta-llama/Llama-3.1-8B-Instruct",
-  "choices": [
-    {
-      "index": 0,
-      "message": {
-        "role": "assistant",
-        "content": "The capital of France is Paris.",
-        "reasoning_content": null,
-        "tool_calls": null
-      },
-      "logprobs": null,
-      "finish_reason": "stop",
-      "matched_stop": 128009
-    }
-  ],
-  "usage": {
-    "prompt_tokens": 12,
-    "total_tokens": 19,
-    "completion_tokens": 7,
-    "prompt_tokens_details": null
-  }
-}
