--- tests/required/client/modelserver/fixtures/sglang/chat_stream.sse
@@ -1,12 +1,0 @@
-data: {"id":"5f2f8f6a1b3c4d5e8f7a6b5c4d3e2f1a","object":"chat.completion.chunk","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","choices":[{"index":0,"delta":{"role":"assistant","content":""},"logprobs":null,"finish_reason":null,"matched_stop":null}]}
-
-data: {"id":"5f2f8f6a1b3c4d5e8f7a6b5c4d3e2f1a","object":"chat.completion.chunk","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","choices":[{"index":0,"delta":{"content":"The capital"},"logprobs":null,"finish_reason":null,"matched_stop":null}]}
-
-data: {"id":"5f2f8f6a1b3c4d5e8f7a6b5c4d3e2f1a","object":"chat.completion.chunk","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","choices":[{"index":0,"delta":{"content":" of France"},"logprobs":null,"finish_reason":null,"matched_stop":null}]}
-
-data: {"id":"5f2f8f6a1b3c4d5e8f7a6b5c4d3e2f1a","object":"chat.completion.chunk","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","choices":[{"index":0,"delta":{"content":" is Paris."},"logprobs":null,"finish_reason":"stop","matched_stop":128009}]}
-
-data: {"id":"5f2f8f6a1b3c4d5e8f7a6b5c4d3e2f1a","object":"chat.completion.chunk","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","choices":[],"usage":{"prompt_tokens":12,"total_tokens":19,"completion_tokens":7,"prompt_tokens_details":null}}
-
-data: [DONE]
-
