--- tests/required/client/modelserver/fixtures/sglang/completion_stream.sse
@@ -1,10 +1,0 @@
-data: {"id":"d3f5a5b8c7e64f2b8a1c9d0e7f6b5a4c","object":"text_completion","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","choices":[{"index":0,"text":" Paris,","logprobs":null,"finish_reason":null,"matched_stop":null}],"usage":null}
-
-data: {"id":"d3f5a5b8c7e64f2b8a1c9d0e7f6b5a4c","object":"text_completion","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","choices":[{"index":0,"text":" the capital","logprobs":null,"finish_reason":null,"matched_stop":null}],"usage":null}
-
-data: {"id":"d3f5a5b8c7e64f2b8a1c9d0e7f6b5a4c","object":"text_completion","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","choices":[{"index":0,"text":" of France.","logprobs":null,"finish_reason":"length","matched_stop":null}],"usage":null}
-
-data: {"id":"d3f5a5b8c7e64f2b8a1c9d0e7f6b5a4c","object":"text_completion","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","choices":[],"usage":{"prompt_tokens":7,"total_tokens":15,"completion_tokens":8,"prompt_tokens_details":null}}
-
-data: [DONE]
-
