--- tests/required/client/modelserver/fixtures/sglang/models.json
@@ -1,13 +1,0 @@
-{
-  "object": "list",
-  "data": [
-    {
-      "id": "meta-llama/Llama-3.1-8B-Instruct",
-      "object": "model",
-      "created": 1750000000,
-      "owned_by": "sglang",
-      "root": "meta-llama/Llama-3.1-8B-Instruct",
-      "max_model_len": 131072
-    }
-  ]
-}
