--- tests/required/client/modelserver/fixtures/tgi/chat.json
@@ -1,23 +1,0 @@
-{
-  "object": "chat.completion",
-  "id": "",
-  "created": 1750000000,
-  "model": "meta-llama/Llama-3.1-8B-Instruct",
-  "system_fingerprint": "3.3.4-native",
-  "choices": [
-    {
-      "index": 0,
-      "message": {
-        "role": "assistant",
-        "content": "The capital of France is Paris."
-      },
-      "logprobs": null,
-      "finish_reason": "stop"
-    }
-  ],
-  "usage": {
-    "prompt_tokens": 12,
-    "completion_tokens": 7,
-    "total_tokens": 19
-  }
-}
