--- tests/required/client/modelserver/fixtures/tgi/completion.json
@@ -1,20 +1,0 @@
-{
-  "object": "text_completion",
-  "id": "",
-  "created": 1750000000,
-  "model": "meta-llama/Llama-3.1-8B-Instruct",
-  "system_fingerprint": "3.3.4-native",
-  "choices": [
-    {
-      "index": 0,
-      "text": " Paris, the capital of France.",
-      "logprobs": null,
-      "finish_reason": "length"
-    }
-  ],
-  "usage": {
-    "prompt_tokens": 7,
-    "completion_tokens": 8,
-    "total_tokens": 15
-  }
-}
