--- tests/required/client/modelserver/fixtures/tgi/completion_stream.sse
@@ -1,10 +1,0 @@
-data: {"object":"text_completion","id":"","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","system_fingerprint":"3.3.4-native","choices":[{"index":0,"text":" Paris,","logprobs":null,"finish_reason":null}],"usage":null}
-
-data: {"object":"text_completion","id":"","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","system_fingerprint":"3.3.4-native","choices":[{"index":0,"text":" the capital","logprobs":null,"finish_reason":null}],"usage":null}
-
-data: {"object":"text_completion","id":"","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","system_fingerprint":"3.3.4-native","choices":[{"index":0,"text":" of France.","logprobs":null,"finish_reason":"length"}],"usage":null}
-
-data: {"object":"text_completion","id":"","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","system_fingerprint":"3.3.4-native","choices":[],"usage":{"prompt_tokens":7,"completion_tokens":8,"total_tokens":15}}
-
-data: [DONE]
-
