--- tests/required/client/modelserver/fixtures/tgi/error.json
@@ -1,4 +1,0 @@
-{
-  "error": "Input validation error: `inputs` tokens + `max_new_tokens` must be <= 4096. Given: 4031 `inputs` tokens and 200 `max_new_tokens`",
-  "error_type": "validation"
-}
