--- tests/required/client/modelserver/fixtures/tgi/models.json
@@ -1,11 +1,0 @@
-{
-  "object": "list",
-  "data": [
-    {
-      "id": "meta-llama/Llama-3.1-8B-Instruct",
-      "object": "model",
-      "created": 0,
-      "owned_by": "meta-llama"
-    }
-  ]
-}
