--- tests/required/client/modelserver/fixtures/vllm/chat_stream.sse
@@ -1,12 +1,0 @@
-data: {"id":"chatcmpl-3c1a9f7e2b3d4e5f8a7b6c5d4e3f2a1b","object":"chat.completion.chunk","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","choices":[{"index":0,"delta":{"role":"assistant","content":""},"logprobs":null,"finish_reason":null}]}
-
-data: {"id":"chatcmpl-3c1a9f7e2b3d4e5f8a7b6c5d4e3f2a1b","object":"chat.completion.chunk","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","choices":[{"index":0,"delta":{"content":"The capital"},"logprobs":null,"finish_reason":null}]}
-
-data: {"id":"chatcmpl-3c1a9f7e2b3d4e5f8a7b6c5d4e3f2a1b","object":"chat.completion.chunk","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","choices":[{"index":0,"delta":{"content":" of France"},"logprobs":null,"finish_reason":null}]}
-
-data: {"id":"chatcmpl-3c1a9f7e2b3d4e5f8a7b6c5d4e3f2a1b","object":"chat.completion.chunk","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","choices":[{"index":0,"delta":{"content":" is Paris."},"logprobs":null,"finish_reason":"stop","stop_reason":null}]}
-
-data: {"id":"chatcmpl-3c1a9f7e2b3d4e5f8a7b6c5d4e3f2a1b","object":"chat.completion.chunk","created":1750000000,"model":"meta-llama/Llama-3.1-8B-Instruct","choices":[],"usage":{"prompt_tokens":12,"total_tokens":19,"completion_tokens":7}}
-
-data: [DONE]
-
