--- tests/required/client/modelserver/fixtures/vllm/models.json
@@ -1,30 +1,0 @@
-{
-  "object": "list",
-  "data": [
-    {
-      "id": "meta-llama/Llama-3.1-8B-Instruct",
-      "object": "model",
-      "created": 1750000000,
-      "owned_by": "vllm",
-      "root": "meta-llama/Llama-3.1-8B-Instruct",
-      "parent": null,
-      "max_model_len": 131072,
-      "permission": [
-        {
-          "id": "modelperm-33f8d4a2b1c04e5f9a6b7c8d9e0f1a2b",
-          "object": "model_permission",
-          "created": 1750000000,
-          "allow_create_engine": false,
-          "allow_sampling": true,
-          "allow_logprobs": true,
-          "allow_search_indices": false,
-          "allow_view": true,
-          "allow_fine_tuning": false,
-          "organization": "*",
-          "group": null,
-          "is_blocking": false
-        }
-      ]
-    }
-  ]
-}
